Organising the repository into folders for robot and ground control code.
