+++ robot-code/default_color_sensor.c
